requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
aiohttp==3.8.5
asyncio-compat==0.1.2
gunicorn==21.2.0
python-dotenv==1.0.0